requirements.txt
Django==3.2.6
djangorestframework==3.14.0
drf-yasg==1.21.6
py2neo==2021.2.3
django-neomodel==0.0.4
python==3.9.6